modified:   B_MNL_Bandit/b_contrastive.py 	modified:   B_MNL_Bandit/b_diagnostics.py 	modified:   B_MNL_Bandit/llm_embedding.py 	modified:   B_MNL_Bandit/mnl_router.py 	modified:   B_MNL_Bandit/queue_config.py 	modified:   B_MNL_Bandit/queue_env.py 	modified:   B_MNL_Bandit/train.py 	renamed:    B_MNL_Bandit/train_by_model/ embedllm_train.py -> B_MNL_Bandit/train_by_model/embedllm_train.py 	modified:   B_MNL_Bandit/train_by_model/mixinstruct_train.py 	new file:   B_MNL_Bandit/train_by_model/routerbench_train.py 	modified:   B_MNL_Bandit/train_by_model/sprout_train.py 	new file:   plot_deperature.py

+++ B_MNL_Bandit/b_diagnostics.py
@@ -180,13 +180,13 @@ def sample_pairs_cos(Z: np.ndarray, max_pairs: int, seed: int) -> np.ndarray:
     rng = np.random.RandomState(seed)
     N = Z.shape[0]
     if N < 2:
-        return np.zeros((0,), dtype=np.float64)
+        return np.zeros((0,), dtype=np.float32)
     P = min(int(max_pairs), N * (N - 1) // 2)
     i = rng.randint(0, N, size=P)
     j = rng.randint(0, N, size=P)
     mask = i != j
     i, j = i[mask], j[mask]
-    return np.sum(Z[i] * Z[j], axis=1).astype(np.float64)
+    return np.sum(Z[i] * Z[j], axis=1).astype(np.float32)
 
 
 @torch.no_grad()
